requirements.txt
pytest==7.3.0
selene==2.0.0rc2
allure-pytest
selenium==4.11.2
webdriver-manager==4.0.0
